requirements.txt
click==3.3 # EL7 version
paramiko==1.16.0
python-neutronclient
python-novaclient
PyYAML>=3.0, <4.0
jinja2
